third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/getTextureFormat-rgba8.https.html

<!DOCTYPE html>
<title>Canvas2D WebGPU Transfer: getTextureFormat-rgba8</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./canvas-test.js"></script>

<h1>Canvas2D WebGPU Transfer: getTextureFormat-rgba8</h1>
<p>getTextureFormat() returns RGBA8 or BGRA8 for a typical context</p>

<script>

runCanvasTestsInWorker();

/** getTextureFormat() should return RGBA8 or BGRA8 for a typical context. */
canvasPromiseTest(async (canvas) => {
  const ctx = canvas.getContext('2d');
  assert_regexp_match(ctx.getTextureFormat(), /^rgba8unorm$|^bgra8unorm$/);
}, 'getTextureFormat() returns RGBA8 or BGRA8 for a typical context');

</script>
